python-ideal-scraper/pyproject.toml
[project]
name = "ideal-scraper"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "requests",
    "beautifulsoup4",
    "sentry-sdk",
    "python-dotenv",
]
